pyproject.toml
[project]
name = "ad-buyer-system"
version = "0.1.0"
description = "CrewAI-based advertising buyer agent using IAB OpenDirect standards"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.121.0",
    "anthropic>=0.40.0",
    "litellm>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[project.scripts]
ad-buyer = "ad_buyer.interfaces.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ad_buyer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
